pyproject.toml
[project]
name = "interactive-fiction-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "attrs==24.3.0",
    "discord-py==2.4.0",
    "frozenlist==1.5.0",
    "idna==3.10",
    "multidict==6.1.0",
    "pexpect==4.9.0",
    "propcache==0.2.1",
    "ptyprocess==0.7.0",
    "python-dotenv==1.0.1",
    "yarl==1.18.3",
]
